Actualizando el instalador

+++ docs/instalador/instalar.sh
@@ -13,7 +13,8 @@ generarConfSphinx()
   sed s/reemplazarID/$(escaparVariable $ID)/g $sources_MERAN/sphinx.conf > /tmp/$ID.sphix.conf
   sed s/reemplazarIUSER/$(escaparVariable  $IUSER_BDD_MERAN)/g /tmp/$ID.sphix.conf > /tmp/$ID.sphix2.conf
   sed s/reemplazarIPASS/$(escaparVariable $IPASS_BDD_MERAN)/g /tmp/$ID.sphix2.conf > /tmp/$ID.sphix.conf
-  sed s/reemplazarDATABASE/$(escaparVariable $BDD_MERAN)/g /tmp/$ID.sphix.conf > $DESTINO_MERAN/$ID/sphinx/etc/sphinx.conf
+  sed s/reemplazarBDDHOST/$(escaparVariable  $HOST_BDD_MERAN)/g /tmp/$ID.sphix.conf > /tmp/$ID.sphix2.conf
+  sed s/reemplazarDATABASE/$(escaparVariable $BDD_MERAN)/g /tmp/$ID.sphix2.conf > $DESTINO_MERAN/$ID/sphinx/etc/sphinx.conf
   rm /tmp/$ID.sphix*
 
 }
@@ -46,10 +47,18 @@ generarJaula()
 actualizarBDD()
 {
    echo "Tenemos que ACTUALIZAR la base de datos $DD_MERAN y para eso pediremos los permisos de root de MySQL"
-
    sed s/reemplazarDATABASE/$(escaparVariable $BDD_MERAN)/g $sources_MERAN/updates.sql > /tmp/UPDATES_MERAN_SANDBOX.sql
-   mysql --default-character-set=utf8  -p < /tmp/UPDATES_MERAN_SANDBOX.sql
-   rm /tmp/UPDATES_MERAN_SANDBOX.sql
+   mysql -h$HOST_BDD_MERAN --default-character-set=utf8 -u$ROOT_USER_BASE --password=$ROOT_PASS_BASE < /tmp/UPDATES_MERAN_SANDBOX.sql
+   if [ $? -ne 0 ]
+	then
+		echo 'No fue posible conecatarse a la base de datos con los datos suministrados'
+		mkdir ~/pendiente
+		mv /tmp/UPDATES_MERAN_SANDBOX.sql ~/pendiente/
+		echo 'Luego debe aplicar el sql que esta en ~/pendiente/UPDATES_MERAN_SANDBOX.sql para actualizar su base de datos'
+	else
+		echo 'Base de datos actualizada con exito'
+	fi
+	rm /tmp/UPDATES_MERAN_SANDBOX.sql
 }
 
 
@@ -62,10 +71,18 @@ generarPermisosBDD()
   tail -n1 $sources_MERAN/permisosbdd.sql | sed s/$(escaparVariable reemplazarDATABASE)/$BDD_MERAN/g > /tmp/$ID.permisosbdd4
   sed s/reemplazarIUSER/$(escaparVariable $IUSER_BDD_MERAN)/g /tmp/$ID.permisosbdd4 > /tmp/$ID.permisosbdd5
   sed s/reemplazarIPASS/$(escaparVariable $IPASS_BDD_MERAN)/g /tmp/$ID.permisosbdd5 >> /tmp/$ID.permisosbdd3
-
+  sed s/reemplazarHOST/$(escaparVariable $DBB_USER_ORIGEN)/g /tmp/$ID.permisosbdd5 >> /tmp/$ID.permisosbdd4  
   echo "Creando la base de Datos..."
-  echo "Tenemos que crear la base de datos $DD_MERAN y para eso pediremos los permisos de root de MySQL"
-  mysql --default-character-set=utf8  -p < /tmp/$ID.permisosbdd3
+  mysql -h$HOST_BDD_MERAN --default-character-set=utf8 -u$ROOT_USER_BASE --password=$ROOT_PASS_BASE < /tmp/$ID.permisosbdd4
+  if [ $? -ne 0 ]
+	then
+		echo 'No fue posible conecatarse a la base de datos con los datos suministrados'
+		mkdir ~/pendiente
+		mv /tmp/$ID.permisosbdd4 ~/pendiente/
+		echo 'Luego debe aplicar el sql que esta en ~/pendiente/'$ID'.permisosbdd4 para crear su base de datos'
+	else
+		echo 'Base de datos creada con exito'
+	fi
   rm /tmp/$ID.permisosbdd*
 }
 
@@ -76,7 +93,8 @@ generarConfiguracion()
   sed s/reemplazarPASS/$(escaparVariable $PASS_BDD_MERAN)/g /tmp/$ID.meran2.conf > /tmp/$ID.meran.conf
   sed s/reemplazarPATHBASE/$(escaparVariable $DESTINO_MERAN)/g  /tmp/$ID.meran.conf > /tmp/$ID.meran2.conf
   sed s/reemplazarCONFMERAN/$(escaparVariable $CONFIGURACION_MERAN)/g /tmp/$ID.meran2.conf > /tmp/$ID.meran3.conf
-  sed s/reemplazarDATABASE/$(escaparVariable $BDD_MERAN)/g /tmp/$ID.meran3.conf > $CONFIGURACION_MERAN/meran$ID.conf
+  sed s/reemplazarBDDHOST/$(escaparVariable  $HOST_BDD_MERAN)/g /tmp/$ID.meran3.conf > /tmp/$ID.meran2.conf
+  sed s/reemplazarDATABASE/$(escaparVariable $BDD_MERAN)/g /tmp/$ID.meran2.conf > $CONFIGURACION_MERAN/meran$ID.conf
   rm /tmp/$ID.meran*
   generarScriptDeInicio
 }
@@ -123,8 +141,6 @@ cambiarPermisos()
       chown www-data:www-data -R $DESTINO_MERAN/$ID/intranet/htdocs/private-uploads
       chown www-data:www-data -R /var/log/meran/$ID/
 
-
-	
 }
 arrancarSphinx()
 {
@@ -147,15 +163,18 @@ usage: $0 options
 Este script necesita si o si el parametro -i que es el identificador de la instalación que se va a utilizar en todo el proceso.
 
 OPTIONS:
-   -h      Show this message
+   -?      Show this message
    -i      Identificador para esta instalación de meran. Este Identificador se va a utilizar en todo
    -d      Carpeta DESTINO donde se guardara meran. Por defecto /usr/share/meran
    -b      Base de datos a usar. Por Defecto va a ser meran
+   -h	   Host donde esta la base de datos
    -u      Usuario que se va a conectar a la base de datos. Por defecto meranadmin
    -p      Pass del usuario que se va a conectar a la base de dato. Por defecto sera un random
    -s      Usuario que se va a utilizar en el indice. Por defecto indice
    -w      Pass del usuario que se va a utilizar en el indice. Por defecto sera un random
    -c      directorio donde se guardará la configuracion de meran. Por defecto sera /etc/meran y el archivo de configuracion sera meran$ID.conf
+   -P	   La pass del usuario para crear la base de datos
+   -U	   El usuario para conectarse a la base de datos
 EOF
 }
 
@@ -174,17 +193,26 @@ USER_BDD_MERAN="kohaadmin"
 PASS_BDD_MERAN=`</dev/urandom tr -dc A-Za-z0-9 | head -c8`
 IUSER_BDD_MERAN="indice"
 IPASS_BDD_MERAN=`</dev/urandom tr -dc A-Za-z0-9 | head -c8`
-
-while getopts “h:i:d:b:u:p:s:w:c:” OPTION
+HOST_BDD_MERAN="localhost"
+ROOT_USER_BASE="root"
+ROOT_PASS_BASE=""
+DBB_USER_ORIGEN="localhost"
+while getopts “?:h:i:d:b:u:p:s:w:c:U:P:” OPTION
 do
      case $OPTION in
-         h)
-             usage
-             exit 1
-             ;;
+         U)
+			ROOT_USER_BASE=$OPTARG
+			;;
+		 P)	
+			ROOT_PASS_BASE=$OPTARG
+            ;;
          i)
              ID=$OPTARG
              ;;
+         h)
+             HOST_BDD_MERAN=$OPTARG
+             DBB_USER_ORIGEN=$(hostname -I)
+             ;;   
          d)
              DESTINO_MERAN=$OPTARG
              ;;
@@ -297,6 +325,12 @@ select OPCION in InstalacionNueva Actualizar
       generarJaula 
       #Crear bdd
       echo "Generando la Base de datos"
+      if [ -z $ROOT_PASS_BASE ]
+		stty -echo
+		echo "Necesitamos un password para acceder al motor con el usuario $ROOT_USER_BASE."
+		read -p "Ingreselo:" ROOT_PASS_BASE
+		stty echo
+	  fi
       generarPermisosBDD
       actualizarBDD
       #Configurar cron
@@ -315,9 +349,15 @@ select OPCION in InstalacionNueva Actualizar
 	   if [ -d $DESTINO_MERAN/$ID ]; 
 	   then
 	     descomprimirArchivos
-      	     cambiarPermisos
-             actualizarBDD
-            /etc/init.d/apache2 restart
+      	 cambiarPermisos
+         if [ -z $ROOT_PASS_BASE ]
+			stty -echo
+			echo "Necesitamos un password para acceder al motor con el usuario $ROOT_USER_BASE."
+			read -p "Ingreselo:" ROOT_PASS_BASE
+			stty echo
+		 fi
+         actualizarBDD
+         /etc/init.d/apache2 restart
 	     break
 	   else
 	     echo "No existe la instalación"
